% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age-adjustment.R
\name{acs_by_geoid}
\alias{acs_by_geoid}
\title{Report ACS values over GEOIDs}
\usage{
acs_by_geoid(input)
}
\arguments{
\item{input}{A Shiny input object}
}
\description{
Report ACS values over GEOIDs
}
