% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age-adjustment.R
\name{acs_sex}
\alias{acs_sex}
\title{Determine which sex to include in the denominator of population calculations}
\usage{
acs_sex(input)
}
\arguments{
\item{input}{A Shiny object, maybe containing filters for sex and/or primary site}
}
\description{
NOTE: THIS MUST BE UPDATED ACCORDING TO THE ROSE TEAM'S PREFERENCES
}
