% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjFit.r
\name{adjFit}
\alias{adjFit}
\title{adjust fitting function}
\usage{
adjFit(bin_info_obj = NULL, bin_count_obj = NULL, sample_id = NULL,
  work_dir = ".", proj_name = "NA", min_n_x = 50,
  trunc_val = 0.999, k_down = 2, k_up = 6, iter_stop = 100,
  reg = "rlm", fdr = 0.05, output_intmd = TRUE)
}
\arguments{
\item{bin_info_obj}{Required. Object of class binInfo, imported using method readBinInfo.}

\item{bin_count_obj}{Required. Object of class binCount, imported using method readBinCount.}

\item{sample_id}{Required. Sample name.}

\item{work_dir}{Working directory. Default is current directory.}

\item{proj_name}{Project name. Default is 'NA'.}

\item{min_n_x}{Minimum number of bins when grouping IP bins with the same Input bin count. Default is 50.}

\item{trunc_val}{Proportion of Input bins not taken as outliers. Default is 0.999.}

\item{k_down}{Minimum k. Default is 2.}

\item{k_up}{Maximum k. Default is 6.}

\item{iter_stop}{Iteration times. Default is 100.}

\item{reg}{Regression method. Valid values are 'rlm' for RLM(robust fitting of linear models) and 'gam' for GAM(generalized additive models).}

\item{fdr}{False discovery rate. Default is 0.05.}

\item{output_intmd}{If save intermediate output. Default is TRUE.}
}
\description{
This is the function for adjusting model fitting.
}
\examples{
#bin_info_*.tsv and bin_count_*.tsv are generated by the main function 'moaims' when setting output_intmd=T(default)
bin_info_obj=readBinInfo(in_fn = /absolute/path/to/bin_info_*.tsv)
bin_count_obj=readBinCount(in_fn = /absolute/path/to/bin_count_*.tsv)
adjFit(bin_info_obj,bin_count_obj,sample_id='test')
}
\author{
Yiqian Zhang
}
