% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age-adjustment.R
\name{age_adj_inc_rate}
\alias{age_adj_inc_rate}
\title{Calculate the age-adjusted incidence rate, with confidence intervals}
\usage{
age_adj_inc_rate(con, input)
}
\arguments{
\item{con}{A database connection (to NEOCASE_Shiny or some such)}

\item{input}{A Shiny input object}
}
\description{
Calculate the age-adjusted incidence rate, with confidence intervals
}
