% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age-adjustment.R
\name{agg_sum}
\alias{agg_sum}
\title{Aggregate and take the sum over a formula. Values are returned as integers.}
\usage{
agg_sum(.frm, .df)
}
\arguments{
\item{.frm}{A formula object, must contain names of columns in `.df`}

\item{.df}{A dataframe to be aggregated}
}
\description{
Aggregate and take the sum over a formula. Values are returned as integers.
}
