% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble-data.R
\name{assemble_data}
\alias{assemble_data}
\title{This function puts together the two datasets that will be presented to the user.}
\usage{
assemble_data(con, input)
}
\arguments{
\item{con}{A database connection (to NEOCASE_Shiny or some such)}

\item{input}{A Shiny input object}
}
\description{
Note: This function is at the front of the call stack, and is thus v important
}
