% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizers.R
\name{clean_nm}
\alias{clean_nm}
\title{Used by 'assets.R' to generate clean filter names}
\usage{
clean_nm(z)
}
\description{
To relabel a filter, include text to replace (first character vector)
  and replacement text (second character vector)
}
\seealso{
selectize_values
}
