% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizers.R
\name{clean_table_names}
\alias{clean_table_names}
\title{Used to clean output table names}
\usage{
clean_table_names(the_table)
}
\arguments{
\item{the_table}{The}
}
\description{
Used to clean output table names
}
\seealso{
assemble_data
}
