% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age-adjustment.R
\name{count_tables}
\alias{count_tables}
\title{Retrieve reference and case counts, with case incidence}
\usage{
count_tables(con, input, agg = FALSE)
}
\arguments{
\item{con}{A database connection (to NEOCASE_Shiny or some such)}

\item{input}{A Shiny input object}

\item{agg}{Whether to aggregate data over GEOID}
}
\description{
Retrieve reference and case counts, with case incidence
}
