% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-helpers.R
\name{csv_nm}
\alias{csv_nm}
\title{Generate a name for a CSV file. Returns a function of no arguments.}
\usage{
csv_nm(f)
}
\description{
Generate a name for a CSV file. Returns a function of no arguments.
}
