% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-acs.R
\name{empty_acs}
\alias{empty_acs}
\title{In the event that ACS calculations can't be returned, this is returned instead}
\usage{
empty_acs(input)
}
\arguments{
\item{input}{A Shiny input object.}
}
\description{
NOTE: Consider rewriting to address 'everyone'
}
