% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age-adjustment.R
\name{everyone_crude_inc}
\alias{everyone_crude_inc}
\title{Get crude incidence for overall population, after filtering by county}
\usage{
everyone_crude_inc(con, input)
}
\arguments{
\item{con}{A database connection (to NEOCASE_Shiny or some such)}

\item{input}{A Shiny input object}
}
\description{
Get crude incidence for overall population, after filtering by county
}
\seealso{
acs_sex agg_sum
}
