% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters-printer.R
\name{filters_printer}
\alias{filters_printer}
\title{This function puts all filter criteria into a neat format for display}
\usage{
filters_printer(input)
}
\arguments{
\item{input}{A Shiny input object}
}
\description{
This function puts all filter criteria into a neat format for display
}
