% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-acs.R
\name{get_acs}
\alias{get_acs}
\title{Retrieve population-level data, calculate relevant proportions}
\usage{
get_acs(con, input)
}
\arguments{
\item{con}{A database connection (to NEOCASE_Shiny or some such)}

\item{input}{A Shiny input object}
}
\description{
Retrieve population-level data, calculate relevant proportions
}
\seealso{
get_ociss empty_acs
}
