% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-input.R
\name{get_filters}
\alias{get_filters}
\title{Retrieve filters from input}
\usage{
get_filters(input)
}
\arguments{
\item{input}{A Shiny input object}
}
\description{
Retrieve filters from input
}
