% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-ociss.R
\name{get_ociss}
\alias{get_ociss}
\title{Retrieve input-appropriate OCISS data}
\usage{
get_ociss(con, input)
}
\arguments{
\item{con}{A database connection (to NEOCASE_Shiny or some such)}

\item{input}{A Shiny input object}
}
\description{
Retrieve input-appropriate OCISS data
}
\seealso{
ociss_sql rename_cc simple_q_gen tt_tx_fn tt_fu_fn ociss_crude_inc age_adj_inc_rate
}
