% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-page-acs.R
\name{mk_acs_page}
\alias{mk_acs_page}
\title{The second page of the interface, this will likely need to be updated as project goals grow}
\usage{
mk_acs_page()
}
\description{
This page has yet to be implemented, the present function is a placeholder.
}
