% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-tab-sidebar-body.R
\name{mk_body}
\alias{mk_body}
\title{Make the body of the app}
\usage{
mk_body(...)
}
\arguments{
\item{...}{Each of the `OCISS` and `a` params from `mk_tab_items`}
}
\description{
Make the body of the app
}
\seealso{
mk_tab_items
}
