% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-page-ociss.R
\name{mk_dat_page}
\alias{mk_dat_page}
\title{Make the user interface page for OCISS table construction}
\usage{
mk_dat_page()
}
\description{
Note that this is where we integrate `filters_ui`, and create the Geographic levels selector (using `ociss::geo_levels`)
}
