% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-tab-sidebar-body.R
\name{mk_sidebar}
\alias{mk_sidebar}
\title{Make the sidebar menu for the interface}
\usage{
mk_sidebar()
}
\description{
Make the sidebar menu for the interface
}
\seealso{
mk_body mk_tab_items
}
