% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-tab-sidebar-body.R
\name{mk_tab_items}
\alias{mk_tab_items}
\title{Set up items for the left-hand side tab in the interface}
\usage{
mk_tab_items(ociss, a)
}
\arguments{
\item{ociss}{The OCISS UI page}

\item{a}{A different UI page (to be implemented)}
}
\description{
Set up items for the left-hand side tab in the interface
}
