% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moaims.r
\name{moaims}
\alias{moaims}
\title{moaims main function}
\usage{
moaims(sample_info_file = NULL, gtf_file = NULL,
  strand_specific = NULL, is_paired = NULL, work_dir = ".",
  proj_name = "NA", bin_size = 200, expr_thred = 0.5,
  per_thred = 0.6, min_n_x = 50, trunc_val = 0.999, k_down = 2,
  k_up = 6, iter_stop = 100, fdr = 0.05, output_intmd = TRUE,
  sep_bin_info = TRUE)
}
\arguments{
\item{sample_info_file}{Required. Sample sheet.}

\item{gtf_file}{Required. Genome annotation in GTF format.}

\item{strand_specific}{Required. Sequencing strand protocol. 0 for unstranded, 1 for fr-first, 2 for fr-second.}

\item{is_paired}{Required. Paired or not.}

\item{work_dir}{Working directory. Default is current directory.}

\item{proj_name}{Project name. Default is 'NA'.}

\item{bin_size}{Bin size. Default is 200.}

\item{expr_thred}{Threshold for expression value(TPM). Default is 0.5.}

\item{per_thred}{At least proportion of the samples listed in the sample sheet express the genes of which bins are generated from. This parameter works when setting sep_bin_info FALSE. Default is 0.6.}

\item{min_n_x}{Minimum number of bins when grouping IP bins with the same Input bin count. Default is 50.}

\item{trunc_val}{Proportion of Input bins not taken as outliers. Default is 0.999.}

\item{k_down}{Minimum k. Default is 2.}

\item{k_up}{Maximum k. Default is 6.}

\item{iter_stop}{Iteration times. Default is 100.}

\item{fdr}{False discovery rate. Default is 0.05.}

\item{output_intmd}{If save intermediate output. Default is TRUE.}

\item{sep_bin_info}{Get bins of genes accoding genes' expression level in the corresponding sample (set TRUE) or in all the samples listed in sample sheet (set FALSE). This parameter is related to per_thred. Default is TRUE.}
}
\description{
This is the main function of moaims, used to detect enriched regions of MeRIP-Seq. Primary outputs include enriched regions(in BED12 format), GOF(Goodness of Fitting) plots and a summary table of models.
}
\examples{
moaims(sample_info_file = /absolute/path/to/sample_info_file, 
	  gtf_file =/absolute/path/to/gtf_file,
	  strand_specific = 1, is_paired = F)
}
\references{
\itemize{
\item Kuan, P.F. et al. A Statistical Framework for the Analysis of ChIP-Seq Data. J Am Stat Assoc 106(495), 891–903 (2011)
\item Bao, Y. et al. Accounting for immunoprecipitation efficiencies in the statistical analysis of ChIP-seq data. BMC Bioinformatics 14, 169 (2013)
}
}
\author{
Yiqian Zhang
}
