% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age-adjustment.R
\name{n_years}
\alias{n_years}
\title{Calculate the number of years that a user has requested}
\usage{
n_years(input)
}
\arguments{
\item{input}{The date range contained in this Shiny input object will be used to calculate the number of years across which data will be reported.}
}
\description{
Calculate the number of years that a user has requested
}
