% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age-adjustment.R
\name{ociss_case_count}
\alias{ociss_case_count}
\title{Get numbers of cases from ociss data, with user input}
\usage{
ociss_case_count(con, input, agg = FALSE)
}
\arguments{
\item{con}{A database connection (to NEOCASE_Shiny or some such)}

\item{input}{A Shiny input object}

\item{agg}{Whether to aggregate data over GEOID}
}
\description{
Get numbers of cases from ociss data, with user input
}
