% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age-adjustment.R
\name{ociss_crude_inc}
\alias{ociss_crude_inc}
\title{Calculate crude incidence}
\usage{
ociss_crude_inc(con, input)
}
\arguments{
\item{con}{A database connection (to NEOCASE_Shiny or some such)}

\item{input}{A Shiny input object}
}
\description{
Calculate crude incidence
}
