% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-ociss.R
\name{ociss_fu_sql}
\alias{ociss_fu_sql}
\title{Generate sql to retrieve time to follow up among user-selected cases}
\usage{
ociss_fu_sql(input)
}
\description{
Generate sql to retrieve time to follow up among user-selected cases
}
