% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-ociss.R
\name{ociss_in}
\alias{ociss_in}
\title{This function shouldn't be called unless there's something to put in 'in' 
Create an 'in' statement from filters and names
This function automatically identifies which lookup table to use}
\usage{
ociss_in(fs, nm)
}
\description{
This function shouldn't be called unless there's something to put in 'in' 
Create an 'in' statement from filters and names
This function automatically identifies which lookup table to use
}
