% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble-data.R
\name{ociss_proportions_by_row}
\alias{ociss_proportions_by_row}
\title{This function calculates the proportions of each column in a table}
\usage{
ociss_proportions_by_row(a_table)
}
\arguments{
\item{a_table}{A table of OCISS counts}
}
\description{
Note: paste/separate are used to keep track of proportions with respect
to their numerators
}
