% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-ociss.R
\name{ociss_tx_sql}
\alias{ociss_tx_sql}
\title{Generate sql to retrieve time to treatment from OCISS data, given user input}
\usage{
ociss_tx_sql(input)
}
\description{
Generate sql to retrieve time to treatment from OCISS data, given user input
}
