% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBinCount.r
\name{readBinCount}
\alias{readBinCount}
\title{import object of class binCount.}
\usage{
readBinCount(in_fn = NULL)
}
\arguments{
\item{in_fn}{Required. Input file bin_count_*, generated by the main function of moaims when setting output_intmd=T(default).}
}
\description{
This is the function for importing object of class binCount.
}
\examples{
bin_count_obj=readBinCount(in_fn = /absolute/path/to/bin_count_*.tsv)
}
\author{
Yiqian Zhang
}
