% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBinInfo.r
\name{readBinInfo}
\alias{readBinInfo}
\title{import object of class binInfo.}
\usage{
readBinInfo(in_fn = NULL)
}
\arguments{
\item{in_fn}{Required. Input file bin_info_*, generated by the main function of moaims when setting output_intmd=T(default).}
}
\description{
This is the function for importing object of class binInfo.
}
\examples{
bin_info_obj=readBinInfo(in_fn = /absolute/path/to/bin_info_*.tsv)
}
\author{
Yiqian Zhang
}
