% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age-adjustment.R
\name{reference_pop}
\alias{reference_pop}
\title{Return reference population counts--these come from the ACS data}
\usage{
reference_pop(input, agg = FALSE)
}
\arguments{
\item{input}{A Shiny input object}
}
\description{
Note that we don't need a DB connection because these data are saved in-package
}
