% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-ociss.R
\name{rename_cc}
\alias{rename_cc}
\title{In a query that returns case counts, rename the 'case_count' column to 
something that is easier to read}
\usage{
rename_cc(df, .nm)
}
\arguments{
\item{df}{A dataframe whose column named 'case_count' will be replaced by `.nm`}

\item{.nm}{The name with which to replace 'case_count' in `df`}
}
\description{
In a query that returns case counts, rename the 'case_count' column to 
something that is easier to read
}
