% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age-adjustment.R
\name{repl_names}
\alias{repl_names}
\title{Convenience function to replace names}
\usage{
repl_names(.df, .repls)
}
\arguments{
\item{.df}{A dataframe whose names are to be replaced}

\item{.repls}{Replacement text, provided as a character vector. The character vector should be named by text to replace (accepting regex), and contents should reflect replacements.}
}
\description{
Convenience function to replace names
}
