% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-helpers.R
\name{safely_in}
\alias{safely_in}
\title{Return several 'in' statements, given a named pairlist of things to have 'in'}
\usage{
safely_in(...)
}
\description{
Return several 'in' statements, given a named pairlist of things to have 'in'
}
