% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-helpers.R
\name{safely_select}
\alias{safely_select}
\title{Create a select clause from a set of groups and/or a first-selected column}
\usage{
safely_select(gps, fst = "count(*) as case_count,")
}
\description{
Create a select clause from a set of groups and/or a first-selected column
}
