% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizers.R
\name{sanitize_filters}
\alias{sanitize_filters}
\title{This function converts filter names to something easier to read. 
See also: ociss/R/filters-printer.R}
\usage{
sanitize_filters(z)
}
\description{
This function converts filter names to something easier to read. 
See also: ociss/R/filters-printer.R
}
