% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-ociss.R
\name{simple_q_gen}
\alias{simple_q_gen}
\title{Use this to generate a simple query, defined as one requiring a single extra string to filter}
\usage{
simple_q_gen(cond, nm)
}
\description{
Use this to generate a simple query, defined as one requiring a single extra string to filter
}
