% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble-data.R
\name{tbl_selector}
\alias{tbl_selector}
\title{This function describes the logic to be used for selecting columns
This is a good place to hide/show anything (e.g. incidence calculations censoring per investigators)}
\usage{
tbl_selector(.df, input)
}
\arguments{
\item{.df}{A table produced by the join of `get_ociss` and `get_acs`}
}
\description{
Note: It is here that we choose to hide Age-Adjusted incidence
}
