% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-helpers.R
\name{text_message}
\alias{text_message}
\title{Send a message. This function is admittedly redundant}
\usage{
text_message(a)
}
\description{
Send a message. This function is admittedly redundant
}
