% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-ociss.R
\name{totals}
\alias{totals}
\title{Retrieve case counts}
\usage{
totals(con, input)
}
\arguments{
\item{con}{A database connection (to NEOCASE_Shiny or some such)}

\item{input}{A Shiny input object}
}
\description{
Retrieve case counts
}
