% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-ociss.R
\name{tt_gen}
\alias{tt_gen}
\title{Create a list of (1) a function to pull data for time to event and (2) the name for that column to be added}
\usage{
tt_gen(.f, .nm)
}
\description{
Create a list of (1) a function to pull data for time to event and (2) the name for that column to be added
}
